+++ test/TypeScript/index.ts
@@ -18,6 +18,8 @@ wav.fromScratch(1, 8000, "32", [[0],[0]]);
 let samples: Float64Array;
 samples = wav.getSamples();
 
+samples = wav.getSamples(true, Int32Array);
+
 try {
 	wav.fromBuffer(new Uint8Array([0]));
 } catch (e) {